backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.0.7
rq==1.16.2
numpy==2.1.3
scikit-learn==1.5.2
rank-bm25==0.2.2
pyyaml==6.0.2
prometheus-client==0.20.0
locust==2.32.4
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-logging==0.48b0

